Rearranges routines from helper to tools
The helper.ts file was cluttered, filled with routines and variables
that belonged to the tools file. This have now been moved.

In addition, the old interface with the language server has been removed
in preparation for the new interface that will be added.
