+++ src/extension.ts
@@ -7,12 +7,12 @@ import FortranHoverProvider from './features/hover-provider';
 import { FortranCompletionProvider } from './features/completion-provider';
 import { FortranDocumentSymbolProvider } from './features/document-symbol-provider';
 
-import { FortranLangServer } from './lang-server';
-import { FORTRAN_DOCUMENT_SELECTOR, EXTENSION_ID, promptForMissingTool } from './lib/helper';
 import { LoggingService } from './services/logging-service';
 import * as pkg from '../package.json';
 import { LANG_SERVER_TOOL_ID } from './lib/tools';
 import { FortranFormattingProvider } from './features/formatting-provider';
+import FortranLanguageServer from './fortls-interface';
+import { EXTENSION_ID, FortranDocumentSelector, promptForMissingTool } from './lib/tools';
 
 // Make it global to catch errors when activation fails
 const loggingService = new LoggingService();
@@ -26,7 +26,7 @@ export function activate(context: vscode.ExtensionContext) {
   if (extensionConfig.get('linterEnabled', true)) {
     const linter = new FortranLintingProvider(loggingService);
     linter.activate(context.subscriptions);
-    vscode.languages.registerCodeActionsProvider(FORTRAN_DOCUMENT_SELECTOR, linter);
+    vscode.languages.registerCodeActionsProvider(FortranDocumentSelector(), linter);
     loggingService.logInfo('Linter is enabled');
   } else {
     loggingService.logInfo('Linter is not enabled');
@@ -34,7 +34,7 @@ export function activate(context: vscode.ExtensionContext) {
 
   if (extensionConfig.get('formatter') !== 'Disabled') {
     const disposable: vscode.Disposable = vscode.languages.registerDocumentFormattingEditProvider(
-      FORTRAN_DOCUMENT_SELECTOR,
+      FortranDocumentSelector(),
       new FortranFormattingProvider(loggingService)
     );
     context.subscriptions.push(disposable);
@@ -45,14 +45,14 @@ export function activate(context: vscode.ExtensionContext) {
 
   if (extensionConfig.get('provideCompletion', true)) {
     const completionProvider = new FortranCompletionProvider(loggingService);
-    vscode.languages.registerCompletionItemProvider(FORTRAN_DOCUMENT_SELECTOR, completionProvider);
+    vscode.languages.registerCompletionItemProvider(FortranDocumentSelector(), completionProvider);
   } else {
     loggingService.logInfo('Completion Provider is not enabled');
   }
 
   if (extensionConfig.get('provideHover', true)) {
     const hoverProvider = new FortranHoverProvider(loggingService);
-    vscode.languages.registerHoverProvider(FORTRAN_DOCUMENT_SELECTOR, hoverProvider);
+    vscode.languages.registerHoverProvider(FortranDocumentSelector(), hoverProvider);
     loggingService.logInfo('Hover Provider is enabled');
   } else {
     loggingService.logInfo('Hover Provider is not enabled');
@@ -60,7 +60,7 @@ export function activate(context: vscode.ExtensionContext) {
 
   if (extensionConfig.get('provideSymbols', true)) {
     const symbolProvider = new FortranDocumentSymbolProvider();
-    vscode.languages.registerDocumentSymbolProvider(FORTRAN_DOCUMENT_SELECTOR, symbolProvider);
+    vscode.languages.registerDocumentSymbolProvider(FortranDocumentSelector(), symbolProvider);
     loggingService.logInfo('Symbol Provider is enabled');
   } else {
     loggingService.logInfo('Symbol Provider is not enabled');
@@ -73,31 +73,5 @@ export function activate(context: vscode.ExtensionContext) {
               For a full list of features the language server adds see:
               https://github.com/hansec/fortran-language-server`;
     promptForMissingTool(LANG_SERVER_TOOL_ID, msg, 'Python', loggingService);
-  }
-
-  // Check that Fortran Intellisense is installed and if not prompt to install
-  if (!vscode.extensions.getExtension('hansec.fortran-ls')) {
-    const msg = `It is highly recommended to install the Fortran IntelliSense 
-              extension. The extension is used to interface with the 
-              fortran-language-server.
-              For a full list of features provided by the extension see:
-              https://github.com/hansec/vscode-fortran-ls`;
-    promptForMissingTool('hansec.fortran-ls', msg, 'VSExt', loggingService);
-  }
-
-  // Our interface with `fortls` has been disabled in favour of the @hansec's
-  // VS Code extension Fortran IntelliSense
-  const useInternalFLInterface = false;
-  if (useInternalFLInterface) {
-    const langServer = new FortranLangServer(context, extensionConfig);
-    langServer.start();
-    langServer.onReady().then(() => {
-      const capabilities = langServer.getCapabilities();
-      if (!capabilities) {
-        return vscode.window.showErrorMessage(
-          'The language server is not able to serve any features at the moment.'
-        );
-      }
-    });
   }
 }
